backend/auth/requirements.txt
pydantic==2.5.0
PyJWT==2.8.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
email-validator==2.1.0
